SDK-4148 remove external strategy
