+++ config/Upgrade/services.yaml
@@ -17,10 +17,6 @@ services:
         bind:
             $isReleaseGroupIntegratorEnabled: '%env(default:is_release_group_integrator_enabled:bool:RELEASE_GROUP_INTEGRATOR_ENABLED)%'
 
-    _instanceof:
-        Upgrade\Application\Strategy\ReleaseApp\Processor\PackageManagerPackagesFetcher\PackageManagerPackagesFetcherInterface:
-            tags: ['upgrage.package_manager_packages_fetcher']
-
     # Upgrade
     Upgrade\:
         resource: '../../src/Upgrade/*'
@@ -329,10 +325,3 @@ services:
             - '%report_send_endpoint_url%'
             - '%report_send_timeout%'
             - '%report_send_connection_timeout%'
-
-    Upgrade\Application\Strategy\ReleaseApp\Processor\ModuleFetcher:
-        arguments:
-            - '@Upgrade\Infrastructure\PackageManager\ComposerAdapter'
-            - '@Upgrade\Application\Strategy\ReleaseApp\Mapper\PackageCollectionMapper'
-            - !tagged_iterator upgrage.package_manager_packages_fetcher
-
